requirements.txt
# Core Flask dependencies
flask==3.0.3
werkzeug==3.0.3
flask-login>=0.6.3
flask-sqlalchemy>=3.1.1
flask-cors>=6.0.1
flask-migrate>=4.1.0

# Database
sqlalchemy>=2.0.0
alembic>=1.17.0

# Google Cloud & Services
google-auth>=2.41.1
google-auth-oauthlib>=1.2.2
google-cloud-secret-manager>=2.25.0
google-cloud-storage>=3.5.0
gspread>=6.2.1

# API Clients
openai>=2.4.0
requests>=2.32.5
requests-oauthlib>=2.0.0
withings-api>=2.4.0
stripe>=13.0.1
twilio>=9.8.0

# HTTP & Networking
httpx>=0.28.0
urllib3>=2.5.0

# Utilities
python-dotenv>=1.1.1
schedule>=1.2.2
pillow>=12.0.0

# Production
gunicorn>=23.0.0
psycopg2-binary>=2.9.11
